requirements.txt
Keras==2.2.4
matplotlib==3.0.2
nltk==3.4
numpy==1.18.5
pandas==0.23.4
pyconll==1.1.3
scikit-learn==0.20.2
scipy==1.5.2
tensorflow-gpu==1.15.0
tqdm==4.50.2
